classify/requirements.txt
bottle==0.12.25
certifi==2023.7.22
charset-normalizer==3.2.0
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
filelock==3.12.3
ftfy==6.1.1
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.25.2
Pillow==10.0.0
regex==2023.8.8
requests==2.31.0
sympy==1.12
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
typing_extensions==4.7.1
urllib3==2.0.4
wcwidth==0.2.6
